server/agent-service/requirements.txt
# LangChain and LangGraph for agent
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
openai>=1.0.0

# FastAPI and web framework
python-dotenv
fastapi
uvicorn[standard]
pydantic
httpx

# MCP for tools
langchain-mcp-adapters>=0.1.0
mcp>=1.0.0
fastmcp>=2.0.0
aiohttp>=3.9.0

# Redis for pub/sub messaging
redis[asyncio]>=5.0.0

# PostgreSQL and SQLAlchemy dependencies (for database access)
sqlalchemy[asyncio]>=2.0.0
asyncpg
psycopg2-binary
